// lib/schedule/models/lesson-type.ts
export enum LessonType {
  Practice = 'practice',
  Lecture = 'lecture',
  LaboratoryWork = 'laboratoryWork',
  IndividualWork = 'individualWork',
  PhysicalEducation = 'physicalEducation',
  Consultation = 'consultation',
  Exam = 'exam',
  Credit = 'credit',
  CourseWork = 'courseWork',
  CourseProject = 'courseProject',
  Unknown = 'unknown'
}